+++ diginode-setup.sh
@@ -10514,13 +10514,6 @@ printf " =============== Checking: DigiNode Tools ==============================
         fi
     fi
 
-    # Ubuntu workaround detected - DigiNode Tools will be installed for the first time
-    if [[ -f "$DGNT_LOCATION/ubuntu-workaround" ]]; then
-        printf "%b %bUbuntu workaround detected. DigiNode Tools will be re-installed to configure aliases etc.%b\\n" "${INFO}" "${COL_LIGHT_GREEN}" "${COL_NC}"
-        DGNT_INSTALL_TYPE="new"
-        DGNT_DO_INSTALL=YES
-    fi
-
     printf "\\n"
 
 }